+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-## Unreleased
+## 0.4.0 (21.2.2017)
 
 - Add some type hints to `metosin.dates`.
 - Remove dangerous keyword -> PGobject `ISQLValue` implementation from `metosin.postgres.types`
 - Removed `ISQLValue` implementation to write maps and vectors as JSON, instead provide new
 `write-json` function to create JSON PGobject.
+- Add `metosin.testing` which provides `assert-expr` method for handling `(is (not ...))`
+- Added `metosin.dates/ToNative` implementation for native date objects, so `to-native`
+returns the same object if the object is already native date object.
+
+**[compare](https://github.com/metosin/metosin-common/compare/0.3.0...0.4.0)**
 
 ## 0.3.0 (11.1.2016)
 
@@ -14,7 +19,7 @@
 - Add `metosin.ring.utils.etag` and `metosin.ring.utils.last-modified` to work with conditional HTTP requests
 - Fix Cache-Control header letter case ([#9](https://github.com/metosin/metosin-common/issues/9))
 
-**[compare](https://github.com/metosin/metosin-common/compare/0.2.3...master)**
+**[compare](https://github.com/metosin/metosin-common/compare/0.2.3...0.3.0)**
 
 ## 0.2.3 (2.8.2016)